+++ HexaEngine.Core/Graphics/ColorWriteEnable.cs
@@ -3,11 +3,11 @@
     [Flags]
     public enum ColorWriteEnable : byte
     {
-        Red = unchecked((byte)1),
-        Green = unchecked((byte)2),
-        Blue = unchecked((byte)4),
-        Alpha = unchecked((byte)8),
-        All = unchecked((byte)15),
-        None = unchecked((byte)0)
+        Red = unchecked(1),
+        Green = unchecked(2),
+        Blue = unchecked(4),
+        Alpha = unchecked(8),
+        All = unchecked(15),
+        None = unchecked(0)
     }
 }
