+++ HexaEngine.Core/Graphics/ShaderResourceViewDimension.cs
@@ -2,17 +2,17 @@
 {
     public enum ShaderResourceViewDimension : int
     {
-        Unknown = unchecked((int)0),
-        Buffer = unchecked((int)1),
-        Texture1D = unchecked((int)2),
-        Texture1DArray = unchecked((int)3),
-        Texture2D = unchecked((int)4),
-        Texture2DArray = unchecked((int)5),
-        Texture2DMultisampled = unchecked((int)6),
-        Texture2DMultisampledArray = unchecked((int)7),
-        Texture3D = unchecked((int)8),
-        TextureCube = unchecked((int)9),
-        TextureCubeArray = unchecked((int)10),
-        BufferExtended = unchecked((int)11)
+        Unknown = unchecked(0),
+        Buffer = unchecked(1),
+        Texture1D = unchecked(2),
+        Texture1DArray = unchecked(3),
+        Texture2D = unchecked(4),
+        Texture2DArray = unchecked(5),
+        Texture2DMultisampled = unchecked(6),
+        Texture2DMultisampledArray = unchecked(7),
+        Texture3D = unchecked(8),
+        TextureCube = unchecked(9),
+        TextureCubeArray = unchecked(10),
+        BufferExtended = unchecked(11)
     }
 }
