+++ HexaEngine.Core/Graphics/UnorderedAccessViewDimension.cs
@@ -2,12 +2,12 @@
 {
     public enum UnorderedAccessViewDimension : int
     {
-        Unknown = unchecked((int)0),
-        Buffer = unchecked((int)1),
-        Texture1D = unchecked((int)2),
-        Texture1DArray = unchecked((int)3),
-        Texture2D = unchecked((int)4),
-        Texture2DArray = unchecked((int)5),
-        Texture3D = unchecked((int)8)
+        Unknown = unchecked(0),
+        Buffer = unchecked(1),
+        Texture1D = unchecked(2),
+        Texture1DArray = unchecked(3),
+        Texture2D = unchecked(4),
+        Texture2DArray = unchecked(5),
+        Texture3D = unchecked(8)
     }
 }
